complete all check and fix bug

+++ UserService/src/main/java/lk/ijse/userservice/service/UserServiceImpl.java
@@ -83,14 +83,18 @@ throw new RuntimeException();
     }
 
 @Override
-public int deleteUser(String email1){
+public int deleteUser(String email1, String password){
       try{
           String email = email1.toLowerCase();
+          String pass = password;
 
           if(userRepository.existsUserByEmail(email)){
               User user = userRepository.findByEmail(email);
-              userRepository.delete(user);
-              return VarList.OK;
+              if(passwordEncoder.matches(pass,user.getPassword())){
+                  userRepository.delete(user);
+                  return VarList.OK;
+              }
+
           }
           return VarList.Not_Found;
       }catch (Exception e){
